requirements.txt
Flask
flask-cors
PyJWT
python-dotenv
